significantly cleaned up my components. now the navbar renders the dialogs instead of the other way around... I'm really disappointed in how long that took me to figure out.
