+++ client/src/components/Dialogs.js
@@ -1,31 +1,7 @@
-// I may end up separating this into two separate components. One dialog for sign in and one for sign up.
-// I didn't at first because of how I'm passing in the Navbar. I don't want to pass too many components into each other.
-
 import React from 'react';
-import Button from '@material-ui/core/Button';
-import TextField from '@material-ui/core/TextField';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-
-// Import Navbar component
-import Navbar from './Navbar'
 
 // Import Function for Storing and Retrieving User's Token
-import {
-  getFromStorage,
-  setInStorage
-} from '../utils/storage';
-
-// Creating some styles to Override material-ui
-const errorStyle = {
-  color: 'red',
-  fontSize: 13,
-  textAlign: 'right',
-  paddingTop: 10
-};
+import { getFromStorage, setInStorage } from '../utils/storage';
 
 // Creates and exports component
 export default class FormDialog extends React.Component {
@@ -65,69 +41,6 @@ export default class FormDialog extends React.Component {
         signedIn: false
       });
     };
-  };
-
-  // Button Handlers
-
-  // Opens Dialog
-  handleClickOpen = () => {
-    this.setState({ open: true });
-  };
-
-  // Closes Dialog
-  handleClose = () => {
-    this.setState({
-      open: false,
-      signInError: '',
-      signUpError: ''
-    });
-  };
-
-  // Changes Dialog From 'Sign In' to 'Sign Up'
-  handleSignedUp = () => {
-    this.setState({signedUp: false});
-  };
-
-  // Handle Submit New User
-  handleUserSubmit = () => {
-    // Target input fields
-    const firstName = document.getElementById('firstName').value;
-    const lastName = document.getElementById('lastName').value;
-    const email = document.getElementById('email').value;
-    const password = document.getElementById('password').value;
-    const passwordVerification = document.getElementById('passwordVerification').value;
-
-    // Posts new user info to DB
-    fetch('api/account/signup', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify({
-        firstName: firstName,
-        lastName: lastName,
-        email: email,
-        password: password,
-        passwordVerification: passwordVerification
-      }),
-    }).then(res => res.json())
-      .then(json => {
-        // If post was successful,
-        if (json.success) {
-          // log out success message
-          this.setState({
-            signUpError: '',
-            signedUp: true,
-            signedIn: true
-          });
-          console.log(`Hello ${firstName}!`);
-        } else {
-        // Else log out error message
-          this.setState({
-            signUpError: json.message
-          });
-        }
-      });
   };
 
   // Handle User Sign In
@@ -202,154 +115,4 @@ export default class FormDialog extends React.Component {
     };
   };
 
-  // Renders Component to app
-  render() {
-    // If User is NOT Signed Up, and NOT signed in
-    // Render navbar with a Sign Up button and a hidden Sign Up dialog
-    // Else if User IS Signed Up, and NOT signed in
-    // Render navbar with a Sign In button and a hidden Sign In dialog
-    // Else Just render Navbar with Sign Out button
-
-    // Sign Up Dialog
-    if (!this.state.signedIn && !this.state.signedUp) {
-      return(
-        <div>
-        {/* Passing the Navbar through the dialog so that the Navbar will be able to open the dialog */}
-        {/* Create Navbar with 'Sign Up' button */}
-        <Navbar open={this.handleClickOpen} isSignedIn={'Sign Up'}/>
-
-        {/* Creates Sign Up Dialog */}
-        <Dialog
-          open={this.state.open}
-          onClose={this.handleClose}
-          aria-labelledby="form-dialog-title"
-          onEnter={this.validateForm}
-        >
-          <DialogTitle id="form-dialog-title">Sign Up</DialogTitle>
-          <DialogContent>
-            <DialogContentText>
-              Please enter the following information
-            </DialogContentText>
-            <TextField
-              autoFocus
-              margin="dense"
-              id="firstName"
-              label="First Name"
-              type="text"
-              fullWidth
-              required
-            />
-            <TextField
-              margin="dense"
-              id="lastName"
-              label="Last Name"
-              type="text"
-              fullWidth
-              required
-            />
-            <TextField
-              margin="dense"
-              id="email"
-              label="Email Address"
-              type="email"
-              fullWidth
-              required
-            />
-            <TextField
-              margin="dense"
-              id="password"
-              label="Password"
-              type="password"
-              fullWidth
-              required
-            />
-            <TextField
-              margin="dense"
-              id="passwordVerification"
-              label="Password Verification"
-              type="password"
-              fullWidth
-              required
-            />
-            <DialogContentText style={errorStyle}>
-              {this.state.signUpError}
-            </DialogContentText>
-          </DialogContent>
-          <DialogActions>
-            <Button onClick={this.handleClose} color="primary">
-              Cancel
-            </Button>
-            <Button onClick={this.handleUserSubmit} color="primary">
-              Submit
-            </Button>
-          </DialogActions>
-        </Dialog>
-        </div>
-      )
-    } 
-
-    // Sign In Dialog
-    else if (!this.state.signedIn && this.state.signedUp) {
-      return (
-        <div>
-          {/* Create Navbar with 'Sign In' button */}
-          <Navbar open={this.handleClickOpen} isSignedIn={'Sign In'}/>
-
-          {/* Creates Sign In dialog*/}
-          <Dialog
-            open={this.state.open}
-            onClose={this.handleClose}
-            aria-labelledby="form-dialog-title"
-          >
-            <DialogTitle id="form-dialog-title">Sign In</DialogTitle>
-            <DialogContent>
-              <DialogContentText>
-                Please enter your user email and password.<br />
-                If you do not have an account please
-                <Button onClick={this.handleSignedUp} color="primary">
-                  Sign Up
-                </Button>
-              </DialogContentText>
-              <TextField
-                autoFocus
-                margin="dense"
-                id="email"
-                label="Email Address"
-                type="email"
-                fullWidth
-              />
-              <TextField
-                margin="dense"
-                id="password"
-                label="Password"
-                type="password"
-                fullWidth
-              />
-              <DialogContentText style={errorStyle}>
-                {this.state.signInError}
-              </DialogContentText>
-            </DialogContent>
-            <DialogActions>
-              <Button onClick={this.handleClose} color="primary">
-                Cancel
-              </Button>
-              <Button onClick={this.handleUserSignIn} color="primary">
-                Sign In
-              </Button>
-            </DialogActions>
-          </Dialog>
-        </div>
-      );
-    } 
-    
-    // Log Out Navbar
-    else {
-      return(
-        <div>
-          {/* Create Navbar with 'Sign Out' button */}
-          <Navbar open={this.handleUserSignOut} isSignedIn={'Sign Out'}/>
-        </div>
-      );
-    }
-  };
 };
